APPDEV-1139 Refactoring the redundant code
-- Moved Redundant code to BasePasscodeActivity.
-- Made BasePasscodeActivity as abstract.
